Added checks on whether support assist/dell commandline updater are already installed or not

+++ scripts/dellupdate.ps1
@@ -1,7 +1,14 @@
-powershell -NoProfile -ExecutionPolicy Bypass -File ".\scripts\check_winget.ps1"
-
-if ( -Not (Test-Path -Path "C:\Program Files (x86)\Dell\CommandUpdate\dcu-cli.exe")){
+if (( -Not (Test-Path -Path "C:\Program Files (x86)\Dell\CommandUpdate\dcu-cli.exe")) -and ( -Not (Test-Path -Path "C:\Program Files\Dell\CommandUpdate\dcu-cli.exe"))) {
+    
+    Write-Host "Installing Dell Commandline." 
+  
     winget install --id Dell.CommandUpdate --accept-package-agreements --accept-source-agreements
 }
-Start-Process "C:\Program Files (x86)\Dell\CommandUpdate\dcu-cli.exe" -ArgumentList "/scan" -Wait
-Start-Process "C:\Program Files (x86)\Dell\CommandUpdate\dcu-cli.exe" -ArgumentList "/applyupdates" -Wait
+
+if ( Test-Path -Path "C:\Program Files\Dell\CommandUpdate\dcu-cli.exe"){
+    Start-Process "C:\Program Files\Dell\CommandUpdate\dcu-cli.exe" -ArgumentList "/scan" -Wait
+    Start-Process "C:\Program Files\Dell\CommandUpdate\dcu-cli.exe" -ArgumentList "/applyupdates" -Wait
+}elseif ( Test-Path -Path "C:\Program Files (x86)\Dell\CommandUpdate\dcu-cli.exe"){
+    Start-Process "C:\Program Files (x86)\Dell\CommandUpdate\dcu-cli.exe" -ArgumentList "/scan" -Wait
+    Start-Process "C:\Program Files (x86)\Dell\CommandUpdate\dcu-cli.exe" -ArgumentList "/applyupdates" -Wait
+}
